Commit a

+++ PI_API/models/leads/Empresa.cs
@@ -6,11 +6,11 @@ public class Empresa
     public string? RazaoSocial { get; set; }
     public string? NaturezaJuridica { get; set; }
     public string? QualificacaoResponsavel { get; set; }
-    public decimal CapitalSocial { get; set; }
+    public double CapitalSocial { get; set; }
     public string? PorteEmpresa { get; set; }
     public string? EnteFederativo { get; set; }
 
-    public Empresa(string? cnpjBase, string? razaoSocial, string? naturezaJuridica, string? qualificacaoResponsavel, decimal capitalSocial, string? porteEmpresa, string? enteFederativo)
+    public Empresa(string? cnpjBase, string? razaoSocial, string? naturezaJuridica, string? qualificacaoResponsavel, double capitalSocial, string? porteEmpresa, string? enteFederativo)
     {
         CnpjBase = cnpjBase;
         RazaoSocial = razaoSocial;
